+++ src/views/wallet/Address.vue
@@ -123,6 +123,25 @@
       <card class="h-100" v-if="!currentAddress">
         <card-header class="card-header fw-bold">
           {{ $route.params.address }}
+          <div class="btn-toolbar" :class="!isMobile ? 'float-end' : 'float-start mt-2'">
+            <div class="btn-group me-2">
+              <button type="button" class="btn btn-outline-theme" @click="copyText($route.params.address)">
+                <i class="fa fa-clipboard hover-info text-secondary ms-1" aria-hidden="true"></i>
+              </button>
+              <button type="button" class="btn btn-outline-theme"
+                      data-bs-toggle="modal"
+                      data-bs-target="#modalQr">
+                <i class="fas fa-lg fa-fw fa-qrcode hover-info text-primary me-1"  aria-hidden="true"></i>
+              </button>
+              <button type="button" class="btn btn-outline-theme"
+                      data-bs-toggle="modal"
+                      data-bs-target="#modalDecryptAddress"
+              >
+                <i class="fas fa-lg fa-fw fa-key hover-info text-warning" @click="decryptSecret()"
+                   aria-hidden="true"></i>
+              </button>
+            </div>
+          </div>
         </card-header>
         <card-body>
           <h3>{{ $t("cold_address") }}</h3>